Move structs used for EIP query to their own file
