+++ crates/cosmos/cosmos-chain-components/src/impls/queries/eip/mod.rs
@@ -1,3 +1,4 @@
 pub mod dispatch;
 pub mod feemarket;
 pub mod osmosis;
+pub mod types;
